moving function to utils
